# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Friction MVP 1.0.0 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles
      parameters:
        - name: authorId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, DIVIDING, CLOSING, LETTER]
      responses:
        "200":
          description: Article list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create a new article (DRAFT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleBody"
      responses:
        "201":
          description: Created article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by ID
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update article content/title
      description: Only allowed when status is not LETTER
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleBody"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          description: Validation error or LETTER immutability violation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete article
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /articles/{id}/transition:
    post:
      operationId: transitionArticleStatus
      tags: [articles]
      summary: Transition article status
      description: "Forward only: DRAFT→DIVIDING→CLOSING→LETTER. LETTER transition sets letter_at and makes article immutable."
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionArticleBody"
      responses:
        "200":
          description: Transitioned article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Inbox ────────────────────────────────────────
  /inbox:
    get:
      operationId: listInbox
      tags: [inbox]
      summary: List inbox items for a user
      description: Only returns items where visible_at <= now
      parameters:
        - name: recipientId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Inbox items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InboxItem"
  /inbox/{id}:
    get:
      operationId: getInboxItem
      tags: [inbox]
      summary: Get inbox item
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Inbox item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInboxItem
      tags: [inbox]
      summary: Delete inbox item
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inbox/{id}/open:
    post:
      operationId: markInboxOpened
      tags: [inbox]
      summary: Mark inbox item as opened (sets opened_at)
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Updated inbox item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inbox/{id}/read:
    post:
      operationId: markInboxRead
      tags: [inbox]
      summary: Mark inbox item as read (完독)
      description: Sets is_read=true and creates UserArticleRead record
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Updated inbox item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── MyCollections ────────────────────────────────
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, email, nickname, createdAt, updatedAt]
...
    Article:
      type: object
      properties:
        id:
          type: string
          format: uuid
        authorId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [DRAFT, DIVIDING, CLOSING, LETTER]
        pages:
          type: array
          items:
            type: string
          nullable: true
        style:
          type: object
          nullable: true
        letterAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, authorId, title, content, status, createdAt, updatedAt]
    CreateArticleBody:
      type: object
      properties:
        authorId:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
        content:
          type: string
      required: [authorId, title]
    UpdateArticleBody:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        pages:
          type: array
          items:
            type: string
          nullable: true
        style:
          type: object
          nullable: true
    TransitionArticleBody:
      type: object
      properties:
        targetStatus:
          type: string
          enum: [DIVIDING, CLOSING, LETTER]
      required: [targetStatus]

    # ─── Inbox ──────────────────────────────────────
    InboxItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        recipientId:
          type: string
          format: uuid
        articleId:
          type: string
          format: uuid
        senderId:
          type: string
          format: uuid
        visibleAt:
          type: string
          format: date-time
        openedAt:
          type: string
          format: date-time
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        article:
          $ref: "#/components/schemas/Article"
        sender:
          $ref: "#/components/schemas/User"
      required: [id, recipientId, articleId, senderId, visibleAt, isRead, createdAt]

    # ─── MyCollection ───────────────────────────────
...