# lib/api-zod/src/generated/types/createArticleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Friction MVP 1.0.0 API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateArticleBody {
  authorId: string;
  /** @minLength 1 */
  title: string;
  content?: string;
}