# lib/api-zod/src/generated/types/inboxItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Friction MVP 1.0.0 API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from "./article";
import type { User } from "./user";

export interface InboxItem {
  id: string;
  recipientId: string;
  articleId: string;
  senderId: string;
  visibleAt: Date;
  openedAt?: Date | null;
  isRead: boolean;
  createdAt: Date;
  article?: Article;
  sender?: User;
}